webapi/tct-extra-html5-tests/semantics/forms/the-textarea-element/textarea_element_attribute.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Xu, YuhanX <yuhanx.xu@intel.com>

-->
<html>
  <head>
    <meta charset="utf-8">
    <title>HTML Test: textarea_element_attribute</title>
    <link rel="author" title="Intel" href="http://www.intel.com/">
    <link rel="help" href="http://www.w3.org/TR/2012/CR-html5-20121217/forms.html#the-textarea-element">
    <script src="../../../resources/testharness.js"></script>
    <script src="../../../resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <textarea id="test_obj"></textarea>
    <script>
        var textarea = document.querySelector("#test_obj");
        [
            ["long", "cols", "20", ""],
            ["HTMLFormElement", "form", "", "readonly"],
            ["long", "rows", "2", ""],
            ["", "type", "textarea", "readonly"],
            ["", "textLength", "", "readonly"],
            ["", "willValidate", "", "readonly"],
            ["ValidityState", "validity", "", "readonly"],
            ["", "validationMessage", "", "readonly"],
            ["NodeList", "labels", "", "readonly"]
        ].forEach(function(attr) {
            var type = attr[0];
            var name = attr[1];
            var defaultValue = attr[2];
            var readOnly = attr[3];

            switch(type) {
                case "long":
                    defaultValue = defaultValue ? parseInt(defaultValue, 10) : defaultValue;
                    break;
                case "HTMLFormElement":
                case "ValidityState":
                case "NodeList":
                    type = "object";
                    break;
                default:
                    break;
            }
            //check type
            if (type == "object") {
                test(function() {
                    assert_equals(typeof textarea[name], type, "Check if the " + name + " is " + type + " type");
                }, "Check if the " + name + " is " + type + " type");
            }

            //check default value
            if (defaultValue) {
                test(function() {
                    assert_equals(textarea[name], defaultValue, "Check if the " + name + " default value is " + defaultValue);
                }, "Check if the " + name + " default value is " + defaultValue);
            }

            //check readonly
            if (readOnly) {
                test(function() {
                    assert_readonly(textarea, name, "Check if the " + name + " attribute is readonly");
                }, "Check if the " + name + " attribute is readonly");
            }
        });
    </script>
  </body>
</html>
